+++ tsup.config.ts
@@ -25,4 +25,16 @@ export default defineConfig([
     splitting: false,
     external: ['react'],
   },
+  // SolidJS integration entry point
+  {
+    entry: ['src/solid/index.ts'],
+    format: ['cjs', 'esm'],
+    dts: true,
+    sourcemap: true,
+    target: 'es2018',
+    outDir: 'dist/solid',
+    treeshake: true,
+    splitting: false,
+    external: ['solid-js'],
+  },
 ]);